+++ assets/controllers/collection-controller.js
@@ -1,5 +1,6 @@
 
 import { Controller } from "@hotwired/stimulus";
+import Swal from 'sweetalert2';
 import { createClient } from '@supabase/supabase-js'
 
 const supabaseUrl = 'URLFORSUPA'
@@ -12,14 +13,30 @@ export default class extends Controller {
 	async deleteImage(event){
 		console.log(event.params);
 		let {id, parent} = event.params;
-		document.getElementById(parent).remove();
-		
-		const { ImageData , ImageError } = await supabase
-			.from('Image')
-			.delete()
-			.eq('imageId', id)
-			.select()
-		console.log("Image: ",ImageData,ImageError);
+		Swal.fire({
+			title: 'Are you sure?',
+			text: "You won't be able to revert this!",
+			icon: 'warning',
+			showCancelButton: true,
+			confirmButtonColor: '#3085d6',
+			cancelButtonColor: '#d33',
+			confirmButtonText: 'Yes, delete it!'
+		}).then(async(result) => {
+		if (result.isConfirmed) {
+			document.getElementById(parent).remove();			
+			const { ImageData , ImageError } = await supabase
+				.from('Image')
+				.delete()
+				.eq('imageId', id)
+				.select()
+			console.log("Image: ",ImageData,ImageError);
+			Swal.fire(
+				'Deleted!',
+				'Your file has been deleted.',
+				'success'
+			)
+		}
+		})
 	}
 
 }
